+++ sniper/setup/__init__.py
@@ -1,0 +1,1 @@
+"""Interactive setup helpers."""